# apps/backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{projectId}/repository": {
      "post": {
        "operationId": "RepositoriesController_createRepository",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "summary": "手动创建仓库（用于未自动创建Repository的旧项目）\nPhase 3: 提供UI按钮让用户初始化Repository",
        "tags": ["Repositories"]
      },
      "get": {
        "operationId": "RepositoriesController_getRepository",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "summary": "获取仓库信息",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/branches": {
      "post": {
        "operationId": "RepositoriesController_createBranch",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBranchDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "summary": "创建分支",
        "tags": ["Repositories"]
      },
      "get": {
        "operationId": "RepositoriesController_getBranches",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "summary": "获取分支列表",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/branches/{branchId}/files": {
      "post": {
        "operationId": "RepositoriesController_uploadFile",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "summary": "上传文件",
        "tags": ["Repositories"]
      },
      "get": {
        "operationId": "RepositoriesController_getFiles",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "summary": "获取文件列表",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/branches/{branchId}/files/download": {
      "get": {
        "operationId": "RepositoriesController_downloadFile",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "summary": "下载文件",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/commits": {
      "post": {
        "operationId": "RepositoriesController_createCommit",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryCreateCommitDto"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "summary": "创建提交",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/branches/{branchId}/commits": {
      "get": {
        "operationId": "RepositoriesController_getCommits",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "summary": "获取提交历史",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/branches/{branchId}/commits/{commitId}": {
      "get": {
        "operationId": "RepositoriesController_getCommit",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "commitId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "summary": "获取提交详情",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/branches/{branchId}/commits/{commitId}/diff": {
      "get": {
        "operationId": "RepositoriesController_getCommitDiff",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "commitId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "compareTo",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "summary": "获取提交间差异",
        "tags": ["Repositories"]
      }
    },
    "/api/projects/{projectId}/repository/branches/{branchId}/commits/{commitId}/files": {
      "get": {
        "operationId": "RepositoriesController_getCommitFiles",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "commitId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "path",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        },
        "summary": "获取提交的文件内容",
        "tags": ["Repositories"]
      }
    },
...
  },
  "info": {
    "title": "Flotilla API",
    "description": "\n基于云计算的开发协作平台 RESTful API 文档\n\n## Rate Limiting 限流策略\n\n本API使用全局限流保护，防止滥用和确保服务可用性。\n\n### 全局限制\n- **默认限制**: 100 requests / minute (所有endpoint)\n\n### 敏感endpoint限制\n以下endpoint有更严格的限流策略：\n- `POST /auth/forgot-password`: **5 requests / hour**\n- `POST /auth/resend-verification`: **5 requests / hour**\n\n### Rate Limit Headers\n每个API响应都包含以下headers：\n- `X-RateLimit-Limit`: 时间窗口内的请求限制数\n- `X-RateLimit-Remaining`: 剩余可用请求数\n- `X-RateLimit-Reset`: 限制重置的Unix时间戳\n\n### 429 Too Many Requests\n当超过限流限制时，API将返回HTTP 429状态码，响应中包含`Retry-After` header指示需要等待的秒数。\n\n**示例响应**:\n```json\n{\n  \"statusCode\": 429,\n  \"message\": \"ThrottlerException: Too Many Requests\"\n}\n```\n    ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "输入 JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateBranchDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "baseBranchId": { "type": "string" }
        },
        "required": ["name"]
      },
      "RepositoryCreateCommitDto": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "minLength": 1, "maxLength": 500 },
          "branchId": { "type": "string" }
        },
        "required": ["message", "branchId"]
      },
...
    }
  }
}